- Blade Angle param chaged to System Load, added potentiometer to simulate such a variable

+++ Core/Inc/nearest_neighbour.h
@@ -9,7 +9,7 @@
 #define INC_NEAREST_NEIGHBOUR_H_
 
 int abs(int);
-int nearest_neighbor(int val, int *arr, int size);
-int find_min_distance(int *arr, int val, int size, int index, int left, int right);
+int nearest_neighbor(int val, uint8_t* arr, int size);
+int find_min_distance(uint8_t* arr, int val, int size, int index, int left, int right);
 
 #endif /* INC_NEAREST_NEIGHBOUR_H_ */
